Introduce Lettuce-based Redis driver
This driver has been created to faciliate usage of the scheduler outside
a Spring environment.

The existing test suite has been re-used (see `AcceptanceTestBase`) for
testing the new driver in integration with a fake Redis server.

Both Spring Data Redis and Lettuce dependencies are now optional, and
must be provided in each projects using the library.